Правки по ДЗ
