+++ index.html
@@ -87,8 +87,7 @@
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--wide">
           <label class="ad-form__label" for="title">Заголовок объявления</label>
-          <input id="title" name="title" type="text" placeholder="Милая, уютная квартирка в центре Токио" minlength="30" maxlength="100" data-pristine-required-message="Обязательное поле" data-pristine-minlength-message="Минимальная длина 30 символов"
-          data-pristine-maxlength-message="Максимальная длина 100 символов" required>
+          <input id="title" name="title" type="text" placeholder="Милая, уютная квартирка в центре Токио" minlength="30" maxlength="100" data-pristine-required-message="Обязательное поле" data-pristine-minlength-message="Минимальная длина 30 символов" data-pristine-maxlength-message="Максимальная длина 100 символов" required>
         </fieldset>
         <fieldset class="ad-form__element ad-form__element--wide">
           <label class="ad-form__label" for="address">Адрес (координаты)</label>
