+++ js/main.js
@@ -2,10 +2,16 @@ import { QUANTITY_OBJECTS } from './data.js';
 import { getNewRandomListings, getListing } from './listing-generator.js';
 import { createOfferTemplate } from './template-card.js';
 import './user-form.js';
-import './page-status.js';
+import { deactivatePage, activatePage } from './page-status.js';
 
 
 const [offer] = getNewRandomListings(QUANTITY_OBJECTS, getListing);
 const offerTemplate = createOfferTemplate(offer);
 
-document.querySelector('#map-canvas').append(offerTemplate);
+const mapCanvas = document.querySelector('#map-canvas');
+
+mapCanvas.append(offerTemplate);
+
+deactivatePage();
+
+mapCanvas.addEventListener('click', () => activatePage());
