+++ js/util.js
@@ -68,26 +68,6 @@ const getRandomProperty = (obj) => {
 
 const isEscapeKeyPressed = (evt) => evt.key === 'Escape';
 
-// Создание высплывающего окна
-
-const createPopup = (template) => {
-  const popup = template.cloneNode(true).querySelector('div');
-  document.body.append(popup);
-  const closePopup = () => {
-    popup.remove();
-    document.removeEventListener('keydown', keyCloseHandler);
-  };
-
-  function keyCloseHandler(evt) {
-    if (isEscapeKeyPressed(evt)) {
-      evt.preventDefault();
-      closePopup();
-    }
-  }
-  popup.addEventListener('click', () => closePopup());
-  document.addEventListener('keydown', keyCloseHandler);
-};
-
 export {
   getArrayRandomElement,
   getRandomNumber,
@@ -96,5 +76,4 @@ export {
   getRandomArrayPart,
   getRandomProperty,
   isEscapeKeyPressed,
-  createPopup
 };
